+++ api_v1/job/views.py
@@ -82,6 +82,10 @@ class JobAPI(APIView):
         if job.status == Job.STATUS['DONE']:
             return Response('Target job has already been done')
 
+        if job.operation not in Job.CAN_CANCEL_OPERATIONS:
+            return Response('Target job cannot be canceled',
+                            status=status.HTTP_400_BAD_REQUEST)
+
         # update job.status to be canceled
         job.update(Job.STATUS['CANCELED'])